requirements.txt
discord.py
openai
pyjokes
aiohttp
homeassistant-api
